+++ python/providers/langgraph/composio_langgraph/provider.py
@@ -59,7 +59,8 @@ class LanggraphProvider(
         )
         action_func.__signature__ = Signature(  # type: ignore
             parameters=get_signature_format_from_schema_params(
-                schema_params=schema_params
+                schema_params=schema_params,
+                skip_default=self.skip_default,
             )
         )
         action_func.__doc__ = description