+++ ts/packages/cli/vitest.config.ts
@@ -28,6 +28,7 @@ export default defineConfig({
     },
   },
   test: {
+    testTimeout: 15_000,
     typecheck: {
       tsconfig: './tsconfig.test.json',
     },